current Version issue
fixed current version issue which was stoping app from show review popup
current Version issue
