+++ Sources/ReviewService/Helper/Preference/ReviewPreferenceManger.swift
@@ -6,6 +6,18 @@
 //
 
 import Foundation
+
+protocol PreferenceDelegator {
+   func getCount(for key: String) -> Int
+   func updateCount(count: Int, using key: String)
+   func getLastVersionPromotedForReview(for key: String) -> String?
+   func updateAppVersion(versionToBeUpdated version: String, using key: String)
+}
+
+protocol AppBundlableDelegator {
+   func getCurrentAppVersion() -> String?
+}
+
 class ReviewPreferenceManger: PreferenceDelegator {
     let userDefaults: UserDefaults
